clean and improve code
