+++ package.json
@@ -1,12 +1,12 @@
 {
   "name": "node.js",
   "version": "0.1.0",
-  "description": "Make CRUD with Node.js, Express backend and MongoDB for database",
+  "description": "Make CRUD with Node.js",
   "main": "index.js",
   "scripts": {
     "start:server": "nodemon ./dist/server.js",
-    "compile": "tsc -p ./tsconfig.json",
-    "build:webpack-dev": "webpack",
+    "watch": "tsc -w -p tsconfig.json",
+    "compile": "tsc -p tsconfig.json",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -24,6 +24,7 @@
     "axios": "^0.19.2",
     "cors": "^2.8.5",
     "express": "^4.17.1",
+    "pg": "^8.2.1",
     "xml-js": "^1.6.11"
   },
   "devDependencies": {
@@ -31,6 +32,7 @@
     "@types/cors": "^2.8.6",
     "@types/express": "^4.17.6",
     "@types/node": "^14.0.4",
+    "@types/pg": "^7.14.3",
     "body-parser": "^1.19.0",
     "nodemon": "^2.0.4",
     "typescript": "^3.9.3"
